# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NyumbaCheck — Nairobi real estate fraud detection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: Search property listings
      parameters:
        - name: neighbourhood
          in: query
          schema:
            type: string
          description: Filter by canonical neighbourhood name
        - name: listingType
          in: query
          schema:
            type: string
            enum: [rent, sale]
        - name: minBedrooms
          in: query
          schema:
            type: integer
        - name: maxBedrooms
          in: query
          schema:
            type: integer
        - name: minPriceKsh
          in: query
          schema:
            type: number
        - name: maxPriceKsh
          in: query
          schema:
            type: number
        - name: platform
          in: query
          schema:
            type: string
          description: Filter by platform slug
        - name: minFraudScore
          in: query
          schema:
            type: number
          description: Only return listings with fraud score above this value
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Listings found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsPage"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing details with fraud score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Agents ─────────────────────────────────────────────────────────────────
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List and search agents
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search by name or phone number
        - name: minReputationScore
          in: query
          schema:
            type: number
        - name: isBlacklisted
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Agents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentsPage"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent profile and reputation score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Market ─────────────────────────────────────────────────────────────────
...
  /market/stats:
    get:
      operationId: getMarketStats
      tags: [market]
      summary: Get market statistics for a neighbourhood
      parameters:
        - name: neighbourhood
          in: query
          required: true
          schema:
            type: string
        - name: listingType
          in: query
          schema:
            type: string
            enum: [rent, sale]
          default: rent
        - name: days
          in: query
          schema:
            type: integer
            default: 30
          description: Look-back period in days
      responses:
        "200":
          description: Market stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketStats"
  /market/trends:
    get:
      operationId: getMarketTrends
      tags: [market]
      summary: Get price trend time series for a neighbourhood
      parameters:
        - name: neighbourhood
          in: query
          required: true
          schema:
            type: string
        - name: listingType
          in: query
          schema:
            type: string
            enum: [rent, sale]
          default: rent
        - name: days
          in: query
          schema:
            type: integer
            default: 90
      responses:
        "200":
          description: Price trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketTrends"

  # ── Alerts ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - message

    # ── Reports ─────────────────────────────────────────────────────────────
...
    FraudSignal:
      type: object
      properties:
        type:
          type: string
        label:
          type: string
        description:
          type: string
        weight:
          type: number
        score:
          type: number
        rawValue:
          nullable: true
          oneOf:
            - type: string
            - type: number
      required:
        - type
        - label
        - description
        - weight
        - score
    FraudScoreSummary:
      type: object
      properties:
        score:
          type: number
          description: Composite fraud score 0–100
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        summary:
          type: string
          description: Human-readable explanation
        signals:
          type: array
          items:
            $ref: "#/components/schemas/FraudSignal"
        platformCountScore:
          type: number
          nullable: true
        priceSpreadScore:
          type: number
          nullable: true
        agentPhoneOverlapScore:
          type: number
          nullable: true
        daysOnMarketScore:
          type: number
          nullable: true
        imageReuseScore:
          type: number
          nullable: true
        priceAnomalyScore:
          type: number
          nullable: true
      required:
        - score
        - riskLevel
        - summary
        - signals
    DuplicateListing:
      type: object
      properties:
        platform:
          type: string
        url:
          type: string
        priceKsh:
          type: number
          nullable: true
        agentPhone:
          type: string
          nullable: true
      required:
        - platform
        - url
...
    ListingSummary:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
        url:
          type: string
        title:
          type: string
          nullable: true
        listingType:
          type: string
          enum: [rent, sale]
        priceKsh:
          type: number
          nullable: true
        bedrooms:
          type: integer
          nullable: true
        bathrooms:
          type: integer
          nullable: true
        sqft:
          type: number
          nullable: true
        neighbourhood:
          type: string
          nullable: true
        fraudScore:
          type: number
          nullable: true
        riskLevel:
          type: string
          nullable: true
          enum: [low, medium, high, critical]
        isActive:
          type: boolean
        firstSeenAt:
          type: string
          format: date-time
      required:
        - id
        - platform
        - url
        - listingType
        - isActive
        - firstSeenAt
    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/ListingSummary"
        - type: object
          properties:
            rawAddress:
              type: string
              nullable: true
            imageUrls:
              type: array
              items:
                type: string
            agentName:
              type: string
              nullable: true
            agentPhone:
              type: string
              nullable: true
            daysOnMarket:
              type: integer
              nullable: true
            fraudScoreDetail:
              nullable: true
              allOf:
                - $ref: "#/components/schemas/FraudScoreSummary"
            duplicates:
              type: array
              items:
                $ref: "#/components/schemas/DuplicateListing"
    ListingsPage:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummary"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - pageSize
        - totalPages

    # ── Agents ──────────────────────────────────────────────────────────────
    AgentSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        reputationScore:
          type: number
          nullable: true
        totalListings:
          type: integer
        ghostListingRate:
          type: number
          nullable: true
        isVerified:
          type: boolean
        isBlacklisted:
          type: boolean
      required:
        - id
        - totalListings
        - isVerified
        - isBlacklisted
    AgentDetail:
      allOf:
        - $ref: "#/components/schemas/AgentSummary"
        - type: object
          properties:
            phoneNumbers:
              type: array
              items:
                type: string
            priceConsistencyScore:
              type: number
              nullable: true
            duplicateListingRate:
              type: number
              nullable: true
            blacklistReason:
              type: string
              nullable: true
            recentListings:
              type: array
              items:
                $ref: "#/components/schemas/ListingSummary"
    AgentsPage:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentSummary"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required:
        - agents
        - total
        - page
        - pageSize
        - totalPages

    # ── Market ──────────────────────────────────────────────────────────────
...
    MarketStats:
      type: object
      properties:
        neighbourhood:
          type: string
        listingType:
          type: string
          enum: [rent, sale]
        period:
          type: string
          description: e.g. "Last 30 days"
        activeListings:
          type: integer
        uniqueProperties:
          type: integer
        duplicateRate:
          type: number
          nullable: true
          description: Proportion of listings that are duplicates (0–1)
        medianPriceKsh:
          type: number
          nullable: true
        medianPricePerSqftKsh:
          type: number
          nullable: true
        p25PriceKsh:
          type: number
          nullable: true
        p75PriceKsh:
          type: number
          nullable: true
        medianDaysOnMarket:
          type: number
          nullable: true
        listingVelocityIndex:
          type: number
          nullable: true
          description: Normalised 0–100. Higher = faster-moving market
        highFraudListings:
          type: integer
        avgFraudScore:
          type: number
          nullable: true
      required:
        - neighbourhood
        - listingType
        - period
        - activeListings
        - uniqueProperties
        - highFraudListings
    MarketTrendPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        medianPriceKsh:
          type: number
          nullable: true
        activeListings:
          type: integer
        newListings:
          type: integer
        medianDaysOnMarket:
          type: number
          nullable: true
      required:
        - date
        - activeListings
        - newListings
    MarketTrends:
      type: object
      properties:
        neighbourhood:
          type: string
        listingType:
          type: string
          enum: [rent, sale]
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/MarketTrendPoint"
      required:
        - neighbourhood
        - listingType
        - dataPoints

    # ── Alerts ──────────────────────────────────────────────────────────────
...